Moved enum_models.py contents to static_definitions
